backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Career Path Advisor - AI-powered career development assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.1.0",
    "langgraph>=1.0.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.18",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
]
